Make Editor code visualize FluidBT structure
Add modifications to Editor code to make it visualize and edit FluidBT
structure instead of TestStructure. The logic of the code is still the
same, but now it references structures from FluidBT. But it assumes that
ITask is a ScriptableObject. Also assumes that BehaviorTree from FluidBT
implements functions such as: DeleteNode and CreateNode

Signed-off-by: André Gustavo Nakagomi Lopez <andregnl@usp.br>